:recycle: [#4] errorboundary layor 분리
