+++ src/router/index.tsx
@@ -1,29 +1,13 @@
-import { ErrorBoundary } from 'react-error-boundary';
-import { createBrowserRouter, Outlet } from 'react-router-dom';
+import { createBrowserRouter } from 'react-router-dom';
 
-import { QueryErrorResetBoundary } from '@tanstack/react-query';
-
-import Profile from '@/pages/Profile';
-import ErrorFallbackComponent from '@components/ErrorFallback';
 import Content from '@pages/Content';
 import Home from '@pages/Home';
 import Login from '@pages/Login';
+import Profile from '@pages/Profile';
 import Test from '@pages/Test';
 
+import ErrorBoundaryLayor from './components/ErrorBoundaryLayor';
 import PATH from './PATH';
-
-const ErrorBoundaryLayor = () => (
-  <QueryErrorResetBoundary>
-    {({ reset }) => (
-      <ErrorBoundary
-        onReset={reset}
-        fallbackRender={({ error }) => <ErrorFallbackComponent error={error} />}
-      >
-        <Outlet />
-      </ErrorBoundary>
-    )}
-  </QueryErrorResetBoundary>
-);
 
 const router = createBrowserRouter(
   [
